wpt_babel/content-security-policy/script-src/scripthash-changed-2.html

<!DOCTYPE html>
<head>
 <title>
  CSP inline script check is done at #prepare-a-script (hash)
 </title>
 <script src="/resources/testharness.js">
 </script>
 <script src="/resources/testharnessreport.js">
 </script>
 <!--
      'log2 += 'scr2 at #prepare-a-script';' => 'sha256-9vE5NuHfEDoLvk3nPZPDX2/mnG+ZwKhpPuwQZwCDGc4=' (blocked)
      'log2 += 'scr2 at #execute-the-script-block';' => 'sha256-3AdhWTFuyxSUPxmqpTJaFRx3R5WNcyGw57lFoj1rTXw=' (allowed)
    -->
 <meta content="script-src 'self' 'nonce-abc' 'sha256-3AdhWTFuyxSUPxmqpTJaFRx3R5WNcyGw57lFoj1rTXw='" http-equiv="Content-Security-Policy"/>
</head>
<!--
  "Should element's inline behavior be blocked by Content Security Policy?"
  is executed at the time of https://html.spec.whatwg.org/C/#prepare-a-script,
  not at https://html.spec.whatwg.org/C/#execute-the-script-block.
  So when innerText is modified after #prepare-a-script, the text BEFORE
  the modification is used for hash check.
-->
<script nonce="abc">
 "use strict";

var log2 = '';
</script>
<!--  Execution order:
  async script is executed
  -> stylesheet is loaded
  -> inline script is executed. -->
<link href="support/empty.css?dummy=2&amp;pipe=trickle(d2)" rel="stylesheet" type="text/css"/>
<script async="" src="support/change-scripthash-before-execute.js?dummy=2&amp;pipe=trickle(d1)">
</script>
<script id="scr2">
 "use strict";

log2 += 'scr2 at #prepare-a-script';
</script>
<script nonce="abc">
 "use strict";

test(function () {
  assert_equals(log2, '');
}, 'scr2.innerText before modification should be blocked');
</script>
